created new app icon to replace default android studio icon
fixed bugs associated with changes in viewLifeCycle state during transitions:
    - text being populated upside down or missing
    - background and text colors missing
    - transitions being frozen mid transition
    - values not being properly loaded into viewmodel

Now when onPause() is called on the gameFragment during a transition it will update the gameState inclusive of the current transition and restart the transition along with the corresponding input of information into the viewmodel when onResume() is called. this avoids the possibility of losing save progress or corrupting of saved files if the game is abruptly left or shut down whether by the user or the system.

Start screen and help dialog will save the current progress of the transition within the fragment so if the fragment is resumed without destroying the views it will continue the transition where it left off and will load all elements properly. if the app is shut down completely it will start again from the beginning regardless of transition progress on shutdown.

created single variable "currentRow" for all of GameFragment's functions to rely on instead of calling getRow() everytime this is needed