*** MAJOR ARCHITECTURE CHANGES ***
All of these changes will help containerize the individual files and make the app overall much easier to test and conform better to the MVVM architecture used in many production quality apps.

--------------------------------------------------------------------

Added Repository classes to improve separation of concerns between the UI and Data layers and to improve the testability of the app.

As the names imply the [GameStateRepository] is for reading and writing to the [gamesState] SharedPreferences, the [SettingsRepository] for the [settings] SharedPreferences
and the [StatsRepository] for the [stats] SharedPreferences.

--------------------------------------------------------------------

Added [StatsViewModel] and scoped it to the [StatsDialog]:
- This separates the StatsDialog from the Data layer and makes the viewModel responsible for fetching the data to be displayed and calculating the size of the bars.
- Since it is scoped to just the StatsDialog it will be created and the variables updated when the dialog is called on then destroyed after the dialog is closed
-Also moved the code to constrain the widths from the fragment to a binding adapter

---------------------------------------------------------------------

Started work on creating tests for the [GameViewModel] and [GameStateRepository]

---------------------------------------------------------------------

Implemented dataBinding in all of the fragments